% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annual_index_expanded_1966to2017}
\alias{annual_index_expanded_1966to2017}
\title{Annual population trend indices for species/regions in the exanded BBS region, 1966-2017}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 491550 rows and 6 columns.}
\source{
\url{https://www.mbr-pwrc.usgs.gov/bbs/inde_1993-2017_expanded.csv}
}
\usage{
annual_index_expanded_1966to2017
}
\description{
Description TBD
}
\keyword{datasets}
