% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbs_obs}
\alias{bbs_obs}
\title{Contains the most recent release of the BBS observations dataset.}
\format{
A data frame with 6797797 rows and 14 columns (as of 2020-09-09).
\describe{
\item{AOU}{Numeric AOU code.}
\item{Count10}{...}
\item{Count20}{...}
\item{Count30}{...}
\item{Count40}{...}
\item{Count50}{...}
\item{CountryNum}{BBS country codes associated with United States-840, Mexico-484, and Canada-124}
\item{Route}{...}
\item{RouteDataID}{...}
\item{RPID}{Route run type}
\item{SpeciesTotal}{...}
\item{State}{The proper name of the US or Mexican state or Canadian province}
\item{StateNum}{BBS state numbers associated with each 'State' and 'CountryNum'}
\item{StopTotal}{Number of individuals of a given species ('AOU') identified by the observer at the end of every 10 stop within the route.}
\item{Year}{Year in which the BBS observations was taken.}
}
}
\source{
\url{https://www.sciencebase.gov/catalog/item/52b1dfa8e4b0d9b325230cd9}
}
\usage{
bbs_obs
}
\description{
Contains the most recent release of the BBS observations dataset.
}
\keyword{datasets}
