% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsAssistant-package.R
\docType{package}
\name{bbsAssistant-package}
\alias{bbsAssistant}
\alias{bbsAssistant-package}
\title{bbsAssistant: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
