% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bbs.R
\name{download_bbs}
\alias{download_bbs}
\title{Download (and optionally subset) USGS Breeding Bird Survey observations data.}
\usage{
download_bbs(
  version = 2020,
  data.dir = here::here("data-raw/"),
  country.names = NULL,
  state.names = NULL,
  overwrite.bbs = NULL,
  overwrite.routes = NULL
)
}
\arguments{
\item{data.dir}{Where to save the 'raw' BBS data. Defaults to directory 'data-raw' in the current working directory.}

\item{state.names}{Vector of state names Default = NULL (all states). See column 'State' in data("region_codes").}

\item{overwrite.bbs}{Logical. Defaults NULL. If TRUE will overwrite existing BBS data in data.dir. NULL will prompt user to force overwrite if .zip files exist in data.dir.}

\item{country.namess}{Vector of country name(s), capitalization irrelevant. One of c(US, USA, United States, U.S.A., U.S., United States of America, CA, Canada, MX, Mexico). Country identities correspond with country codes (BBS): 484==Mexico; 124==Canada; 840==United States. State/region files DNE for Mexico as of November 2019}
}
\value{
.csv
}
\description{
This function downloads a select subset or all the compressed, state-level BBS data from the permanent store at the USGS ScienceBase repository.
}
