% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bbs_data.R
\name{download_bbs_data}
\alias{download_bbs_data}
\title{Download the BBS dataset files associated with a single ScienceBase identifier.}
\usage{
download_bbs_data(sb_id = NULL, bbs_dir = NULL, overwrite = FALSE)
}
\arguments{
\item{sb_id}{Alphanumeric character. Equals the USGS Science Base identifier for a BBS dataset release.}

\item{bbs_dir}{Directory for storing the downloaded files for sb_id}

\item{overwrite}{Logical, default=TRUE. TRUE will overwrite existing files for sb_id if already exists within bbs_dir}
}
\description{
This is functionally sbtools::item_file_download, but forces the overwrite_file=TRUE condition, and specifies the directory to store item files.
}
