% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_bbsFeathers.R
\name{export_bbsFeathers}
\alias{export_bbsFeathers}
\title{Save BBS dataframe as a feather file on disk.}
\usage{
export_bbsFeathers(dataIn, newDir = here::here("bbsData/"), filename)
}
\arguments{
\item{dataIn}{Data to be saved to local disk (e.g., /bbsData/). This data frame is a product of merging route information and raw count data. Data frame should have at least the columns dplyr::select(year, countrynum, statenum, route, bcr, latitude, longitude,aou, stoptotal).}

\item{newDir}{Where to save the BBS data as feathers. If not specified will default to /bbsData/.}

\item{filename}{Name of the new filename (e.g., 'arizona.zip'). This should be the same filename used to import the .zip from FTP (i.e, it should end with ".zip") This function will replace .zip with .feather}
}
\description{
Save the BBS data frame as a feather file to disk for easier import.
}
