% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMass.R
\name{funcMass}
\alias{funcMass}
\title{Load functional trait and mass data}
\usage{
funcMass(dataWD = paste0(getwd(), "/data"), fxn = TRUE, mass = TRUE)
}
\arguments{
\item{dataWD}{Where the functional trait and mass dataframes are stored.}

\item{fxn}{Logical. Retrieves functional trait data (referece).}

\item{mass}{Logical. Retrieves body mass information (Dunning reference).}
}
\description{
Load functional trait and mass data
}
