% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bbs_data.R
\name{get_bbs_data}
\alias{get_bbs_data}
\title{One-stop shop for downloading and importing the USGS Breeding Bird Survey observations data.}
\usage{
get_bbs_data(sb_id, bbs_version, dir, country = NULL, state = NULL)
}
\arguments{
\item{sb_id}{Specify the dataset to download based on the unique ScienceBase Item identifier for retrieval from ScienceBase. If \code{bbs_version} is specified, this will be ignored.}

\item{bbs_version}{Specify the dataset release version (by release year) for retrieval from ScienceBase. Please see /data-raw/sb_items for a list of available datasets and their associated release year.}

\item{country}{Vector of country name(s), capitalization irrelevant. One of c(US, USA, United States, U.S.A., U.S., United States of America, CA, Canada, MX, Mexico).}

\item{state}{Vector of state names Default = NULL (all states). See column 'State' in data("region_codes").}

\item{data_dir}{Directory within which the BBS compressed and decompressed files associated with the ScienceBase item will be stored. Location defaults to a new folder named after the ScienceBase identifier tag (sb_id), in ~/data-raw/<SB_ID>.}

\item{sb_dir}{Where to save the files associated with the ScienceBase item. efaults to (a created) folder within folder 'data-raw' in the current working directory.}
}
\description{
This function downloads all files assocaited with a single version (defaults to the most recent release) of the BBS observations dataset from USGS ScienceBase repository, and allows the user to import all or a subset (based on Country, State, or Province).
}
