% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bbsAous.R
\name{get_bbsAous}
\alias{get_bbsAous}
\title{Import the species list used by the BBS from website or local disk.}
\usage{

  get_bbsAous(file = "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/SpeciesList.txt",
  skip = 7)
}
\arguments{
\item{file}{A character string for location for the species list (as .txt).}

\item{skipEmpty}{Numeric, default = 7. The number of lines to skip. This may need to be updated if the file on webpage changes.}
}
\description{
This function imports the species list (incuding AOU numbers) from their website. This list can be used to subet the species by name, AOU number, etc.
}
