% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credibility_trends.R
\name{get_credibility_trends}
\alias{get_credibility_trends}
\title{Scrape BBS data credibility scores and population trend estimates from BBS website}
\usage{
get_credibility_trends(url = NULL)
}
\arguments{
\item{url}{The url location of the web page to be scraped. Currently, you will have to specify this url. Default url is House Sparrows in Florida.}
}
\description{
A function using web scraping tools to retrieve region-specific species trend estimates \link{@sauer2017north} and data credibility ratings. Current functionality does not include retrieving estimates per species, but rather per region.
}
\references{
Regional Credibility Measures information can be found online at \url{https://www.mbr-pwrc.usgs.gov/bbs/credhm09.html}
}
