% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_routes.R
\name{get_routes}
\alias{get_routes}
\title{Download and import the route-level information via FTP (USGS BBS)}
\usage{
get_routes(
  routes.url = "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/routes.zip",
  data.dir = here::here("raw-data/"),
  active.only = TRUE,
  route.type = 1,
  bbs.stratum = NULL,
  bcr.num = NULL,
  overwrite = NULL
)
}
\arguments{
\item{routes.url}{URL for location of the routes.zip folder.}

\item{data.dir}{Location for storing the routes.zip and unzipped file(s). We recommend storing this in the raw-data directory.}

\item{active.only}{Logical. Default=TRUE. Keep only active routes (ie.. Active==1).}

\item{route.type}{Default = 1 (roadside surveys only). One or more numbers indicating route substrate (1=roadside;2=water;3=off-road; Default = 1, roadside only).}

\item{bbs.stratum}{Vector of BBS physiographic stratum codes by which to filter the routes.}

\item{bcr.num}{A vector of Bird Conservation Region codes where by which to filter the routes.}

\item{overwrite}{Logical. Default NULL. Will force overwrite existing file named "routes.csv" if it exists in data.dir. If NULL will prompt user to specify overwrite.}
}
\value{
A dataframe containing route-level information.
}
\description{
This function downloads information about route location from the BBS FTP server.
}
\examples{
# download and import the BBS route information
\dontrun{
routes <- get_routes()
}

}
