% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_usgs_results.R
\name{get_sb_items}
\alias{get_sb_items}
\title{Download and Import the Most Recent USGS BBS Analytical Results}
\usage{
get_sb_items()
}
\description{
Downloads and imports the analytical results published on ScienceBase.
Refers to internal dataset \code{sb_items} to capture the most recent ScienceBase identifier, sb_id,
}
