% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unzip.R
\name{get_unzip}
\alias{get_unzip}
\title{Downloads and unzips a zip archive to file.}
\usage{
get_unzip(ZipName, FileName)
}
\arguments{
\item{ZipName}{file to download}

\item{FileName}{file to unzip to}
}
\value{
A dataframe
}
\description{
Downloads and unzips a zip archive to file.
}
\details{
Used internally. If ZipName begins with 'http' or 'ftp', then download and unzip to
Filename and return as a dataframe. Otherwise, unzip Zipname and return as a data.frame.
}
