% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_usgs_results.R
\name{grab_usgs_results}
\alias{grab_usgs_results}
\title{Download and Import the Most Recent USGS BBS Analytical Results}
\usage{
grab_usgs_results()
}
\description{
Downloads and imports the analytical results published on ScienceBase.
Refers to internal dataset \code{sb_items} to capture the most recent ScienceBase identifier, sb_id,
}
