% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_bbs.R
\name{import_bbs}
\alias{import_bbs}
\title{Import BBS data into R from file.}
\usage{
import_bbs(data.dir = paste0(getwd(), "/raw-data/"),
  state.names = NULL)
}
\arguments{
\item{data.dir}{Location of the compressed (.zip) route-level files (one per region/states) to be imported into R. Files will also be unzipped to this directory as ".csv".}

\item{state.names}{Vector of state names Default = NULL (all states). See column 'State' in data("region_codes").}
}
\description{
Import compressed (.zip) BBS data from file, combine all regions/states into a single data.frame object and load into environment.
}
