% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_bbsData.R
\name{import_bbsData}
\alias{import_bbsData}
\title{A wrapper function for downloading raw data, filtering by species names/AOU codes, and merging with route information.}
\usage{
import_bbsData(file, dir, year = NULL, aou = NULL, countrynum = NULL,
  states = NULL, routesFile = "routes.zip",
  routesDir = "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/",
  RouteTypeID = 1, Stratum = NULL, BCR = NULL)
}
\arguments{
\item{file}{The name of the zipfile to be downloaded from dir}

\item{dir}{http:// pr ftp:/ link to bbs data location (here: state files)}

\item{year}{Vector of years. Default = NULL (all years).}

\item{aou}{Vector of AOU #s Default = NULL (all species).}

\item{countrynum}{Vector of country ID #'s. Default = NULL (all countryNums).}

\item{states}{Vector of state names Default = NULL (all states).}

\item{RouteTypeID}{One or more numbers indicating route substrate (1=roadside;2=water;3=off-road; Default = 1, roadside only).}

\item{Stratum}{A vector of BBS physiographic stratum codes by which to filter the routes.}

\item{BCR}{A vector of Bird Conservation Region codes where by which to filter the routes.}

\item{zipFileNames}{One or more file names ("state.zip"), char vector. Preferably download a single state at a time, otherwise run time will take >1 minutes.}

\item{countryName}{Vector of country names. Default = NULL (all countryNames).}

\item{regions}{Vector of regionCodes. Default = NULL (all regions).}

\item{routeDir}{Location of the routes.zip folder Should be in DatFiles folder (default).}

\item{routeFile}{Name of the route information file. Usually "routes.zip".}
}
\description{
This function will import BBS data which has already been downloaded to file. Some parts of this function were borrowed from \strong{oharar/rBBS}.
}
