% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_bbsFeathers.R
\name{import_bbsFeathers}
\alias{import_bbsFeathers}
\title{Load the BBS data feathers into R.}
\usage{
import_bbsFeathers(newDir, filename)
}
\arguments{
\item{newDir}{Where the BBS feathers are saved.}

\item{filename}{Name of the feather filename (e.g., 'arizona.zip' or 'arizona'). This function will replace .zip with .feather when necessary.}
}
\description{
Load the BBS data feathers into R.
}
