% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_routes.R
\name{import_routes}
\alias{import_routes}
\title{Import the route-level covariates/metadata associated with the BBS dataset.}
\usage{
import_routes(sb_dir)
}
\arguments{
\item{sb_dir}{Directory for the ScienceBase (sb) item.}
}
\description{
Unzips and imports the routes.zip (routes.csv) file associated with each BBS dataset.
}
