% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_species_list.R
\name{import_species_list}
\alias{import_species_list}
\title{Import Species Metadata}
\usage{
import_species_list(bbs_dir)
}
\arguments{
\item{bbs_dir}{Directory for the ScienceBase (sb) item.}
}
\description{
Imports the file specieslist.txt. Is subsequently appended to the BBS dataset in \code{import_bbs_data()}
}
