% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_vehicle_data.R
\name{import_vehicle_data}
\alias{import_vehicle_data}
\title{Import the VehicleData (number of vehicles observed during each stop).}
\usage{
import_vehicle_data(sb_dir)
}
\arguments{
\item{sb_dir}{Directory for the ScienceBase (sb) item.}
}
\description{
Unzips and imports the VehicleData.zip (into weather.csv) file associated with each BBS dataset.
}
