% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_weather.R
\name{import_weather}
\alias{import_weather}
\title{Import the weather conditions associated with route-level observations.}
\usage{
import_weather(sb_dir)
}
\arguments{
\item{sb_dir}{Directory for the ScienceBase (sb) item.}
}
\description{
Unzips and imports the Weather.zip (into weather.csv) file associated with each BBS dataset.
}
