% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-make.dates.R
\name{make.dates}
\alias{make.dates}
\title{Make Date and Julian Day Variable}
\usage{
make.dates(df, collapse=FALSE)
}
\arguments{
\item{df}{a flat data object.}

\item{collapse}{logical. If TRUE will remove the columns 'Day', 'Month', and 'Year'}
}
\description{
Creates a new variables, \code{c("Date","julian")}, if dataframe has columns \code{c("Day", "Month", "Year")}
}
