% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.rteno.R
\name{make.rteno}
\alias{make.rteno}
\title{Make Identifying Variable for Routes}
\usage{
make.rteno(df)
}
\arguments{
\item{df}{a flat data object.}
}
\description{
Creates a new variable, \code{RTENO}, if dataframe has columns \code{c("CountryNum", "StateNum", "Route")}
}
