% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_bbs_data.R
\name{munge_bbs_data}
\alias{munge_bbs_data}
\title{Subset BBS observations data by one or more species of interest.}
\usage{
munge_bbs_data(
  bbs_list,
  states = NULL,
  countries = NULL,
  species = NULL,
  genera = NULL,
  order = NULL,
  family = NULL,
  zero.fill = TRUE,
  year.range = 1966:lubridate::year(Sys.Date()),
  keep.stop.level.data = FALSE,
  QualityCurrentID = 1,
  active.only = FALSE,
  observations.output = "flat"
)
}
\arguments{
\item{species}{A vector of one or more species (using English Common Name) to subset the data by. Capitalization ignored.}

\item{genera}{vector of one or more genera  to retain in data}

\item{order}{vector of one or more orders to retain in data}

\item{family}{vector of one or more families to retain in data}

\item{zero.fill}{If TRUE and a single species is provided in 'species', this function will output list$observations with zero-filled data.}

\item{year.range}{value(s) of year(s) by which to subset the data. If not specified will return all data in the dataset.}

\item{active.only}{Logical. If TRUE keep only active routes. Discontinued routes will be discarded.}

\item{observations.output}{return the object as a list of data frames ("list") or a single data frame ("flat" or "df" or "data.frame")}

\item{list}{A list with element "species_list", obtained from running bbsAssistant::get_bbs_data()...}
}
\description{
...
}
