% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_get_bbs.R
\name{quick_get_bbs}
\alias{quick_get_bbs}
\title{A wrapper function for downloading (\code{bbsAssistant::download_bbs()}) from the USGS FTP server and importing (\code{bbsAssistant::import_bbs()}) the BBS into the environment.}
\usage{

  quick_get_bbs(data.link = "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/States/",
  data.dir = here::here("raw-data/"), country.names = NULL,
  state.names = NULL)
}
\arguments{
\item{data.link}{URL to the location of 'States' compressed files on the USGS server. Defaults to the FTP location.}

\item{data.dir}{Where to save the 'raw' BBS data. Defaults to subdir 'raw-data' in the current working directory.}

\item{state.names}{Vector of state names to be downloaded and imported (using functions \code{bbsAssistant::downoad_bbs()} and \code{bbsAssistant::import_bbs()}. Default = NULL (all states). See column 'State' in data("region_codes").}

\item{country.namess}{Vector of country name(s), capitalization irrelevant. One of c(US, USA, United States, U.S.A., U.S., United States of America, CA, Canada, MX, Mexico). Country identities correspond with country codes (BBS): 484==Mexico; 124==Canada; 840==United States. State/region files DNE for Mexico as of November 2019}
}
\description{
This function wraps the functions \code{bbsAssistant::download_bbs()} and \code{bbsAssistant::import_bbs()} for speedy retrieval and imortation of all or a selection of the BBS point count data into the local environment
}
