% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_codes}
\alias{region_codes}
\title{Contains country and state numbers, and the associated zip filenames.#'}
\format{Describe the format of df:
\describe{
\item{CountryNum}{BBS country codes associated with United States (484), Mexico (484), and Canada (840)}
\item{StateNum}{BBS state numbers associated with each 'State' and 'CountryNum'}
\item{State}{The proper name of the (US/MEX) state or (CAN) province}
\item{zip_states}{The filename of the .zip compressed BBS count data located on the FTP server. These files do not exist for Mexican states, so == NA}
}}
\source{
\url{ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/RegionCodes.txt}
}
\usage{
region_codes
}
\description{
Contains country and state numbers, and the associated zip filenames.#'
}
\keyword{datasets}
