% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sb_items}
\alias{sb_items}
\title{A lookup table that is manually updated by the package maintainers. It is used as a quick reference to identify which datasets are available for import from ScienceBase, and are compatible with this package.}
\format{
A data frame containing N rows and 5 columns
\describe{
\item{sb_parent}{ScienceBase unique identifier associated with the parent item (often the project identifier).}
\item{sb_item}{ScienceBase unique identifier associated with each dataset release.}
\item{sb_title}{Title of the item as retrieved from ScienceBase.}
\item{release_year}{Year of the dataset release.}
\item{data_type}{One of c("observations", "results"). ScienceBase items are assigned one of the two to distinguish between the analysis results (results) and the original observations dataset release (observations).}
\item{year_start}{Earliest year of data in the associated dataset.}
\item{year_end}{Earliest year of data in the associated dataset.}
\item{legacy_format}{Used to specify whether the observations dataset is in a different format than more recent releases. This is defined by maintainers but based off of ScienceBase documentation.}
}
}
\usage{
sb_items
}
\description{
A lookup table that is manually updated by the package maintainers. It is used as a quick reference to identify which datasets are available for import from ScienceBase, and are compatible with this package.
}
\keyword{datasets}
