% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_list}
\alias{species_list}
\title{Taxonomic information from teh American Ornithological Society's AOU code.}
\format{
A data frame with 2158 rows and 17 variables.
\describe{
\item{AOU}{Numeric AOU code.}
}
}
\source{
Most recent version of AOU at \url{http://checklist.aou.org/taxa.csv?type=charset\%3Dutf-8\%3Bsubspecies\%3Dno\%3B}

Most recent version of species list associated with BBS data at SpeciesList.txt at \url{http://checklist.aou.org/taxa.csv?type=charset\%3Dutf-8\%3Bsubspecies\%3Dno\%3B"}
}
\usage{
species_list
}
\description{
Taxonomic information from teh American Ornithological Society's AOU code.
}
\keyword{datasets}
