% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_speciesList.R
\name{subset_speciesList}
\alias{subset_speciesList}
\title{Subset the BBS data by species, functional traits, and/or body mass. Default is null.}
\usage{
subset_speciesList(myData, subset.by = NULL, aou.ind = NULL,
  order.ind = NULL, fam.ind = NULL)
}
\arguments{
\item{myData}{A data frame containing the BBS data. Must contain the column "aou".}

\item{aou.ind}{Numeric or vector of numeric values of the AOU codes. These are the species you want to REMOVE from analysis.}

\item{order.ind}{Character or vector of characters of taxonomic orders to remove}

\item{fam.ind}{Character or vector of characters of taxonomic family to remove}
}
\description{
This function allows the user to subset the BBS data by specific species using AOU numbers.
}
