% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_bbs_data.R
\name{unpack_bbs_data}
\alias{unpack_bbs_data}
\title{Decompress the BBS observation files as obtained from ScienceBase.}
\usage{
unpack_bbs_data(sb_dir, country, state)
}
\arguments{
\item{sb_dir}{...}

\item{country}{...}

\item{state}{...}
}
\description{
Decompress the BBS observation files as obtained from ScienceBase.
}
