% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annual_index_core_1966to2017}
\alias{annual_index_core_1966to2017}
\title{Annual population trend indices for species/regions in the core BBS region, 1966-2017}
\format{
A data frame:
\describe{
\item{AOU}{Five digit species identification number according to the bird species checklist of the American Ornithologist Union.}
\item{Region}{Regions are states, Provinces, Bird Conservation Regions, BBS Regions, Canada excluding Newfoundland and Yukon, the lower 48 US states, and survey-wide excluding Newfoundland and Alaska.}
\item{Year}{Four digit year of survey, from 1966-2017}
\item{Index}{Annual index of species/region abundance estimated using a hierarchical model}
\item{2.5\\% CI}{Lower 95\\% confidence interval of the annual index of abundance}
\item{97.5\\% CI}{Upper 95\\% confidence interval of the annual index of abundance}
}
}
\source{
\url{https://www.mbr-pwrc.usgs.gov/bbs/inde_1966-2017_core_v2.csv}
}
\usage{
annual_index_core_1966to2017
}
\description{
Annual population trend indices for species/regions in the core BBS region, 1966-2017
}
\keyword{datasets}
