% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsAssistant-package.R
\docType{package}
\name{bbsAssistant-package}
\alias{bbsAssistant}
\alias{bbsAssistant-package}
\title{bbsAssistant: What the Package Does (Title Case)}
\description{
To learn more about dplyr, start with the vignettes:
\code{browseVignettes(package = "dplyr")}
}
\keyword{internal}
