% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bbs_data.R
\name{download_bbs_data}
\alias{download_bbs_data}
\title{Download the BBS dataset files associated with a single ScienceBase identifier}
\usage{
download_bbs_data(sb_id, sb_dir)
}
\arguments{
\item{sb_id}{Temp description}

\item{sb_dir}{Temp description}
}
\description{
Called inside \code{get_bbs_data()}. This is functionally sbtools::item_file_download, but forces the overwrite_file=TRUE condition, and specifies the directory to store item files.
}
