% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_bbsFeathers.R
\name{export_bbsFeathers}
\alias{export_bbsFeathers}
\title{Save BBS dataframe as a feather file on disk.}
\usage{
export_bbsFeathers(dataIn, newDir)
}
\arguments{
\item{dataIn}{The BBS data to save.}

\item{newDir}{Where to save the BBS data as feathers.}

\item{filename}{Name of the new filename (e.g., 'arizona.zip'). This function will replace .zip with .feather}
}
\description{
Save the BBS data frame as a feather file to disk for easier import.
}
