% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bbsData.R
\name{get_bbsData}
\alias{get_bbsData}
\title{Download USGS Breeding Bird Survey data to file and import into the environment.}
\usage{
get_bbsData(file,
  dir = "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/States/",
  year = NULL, aou = NULL, countrynum = NULL, states = NULL)
}
\arguments{
\item{file}{One file name including the .zip extension ("stateX.zip"). Preferably download a single state at a time, otherwise run time will take >1 minutes.}

\item{dir}{URL to the StatesFiles.}

\item{year}{Vector of years. Default = NULL (all years).}

\item{aou}{Vector of AOU numeric codes. Default = NULL (all species). (For species list visit the BBS FTP site (ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/).}

\item{countrynum}{Vector of country ID #'s. Default = NULL (all countryNums).}

\item{states}{Vector of state names Default = NULL (all states).}
}
\value{
If download successful, a dataframe with the results.
}
\description{
This function downloads a select subset or all the BBS data from the FTP server using the .txt file downloaded from get_regions() or using specified regions (e.g. Florida, Florida.zip). The data are saved to a temporary folder. This function was adapted from \strong{oharar/rBBS} package. Note: this function requires an internet connection. If the bbs data are on file, please specify in the parameter \code{file}.
}
\examples{
# download all species and years from Nebraska.

\dontrun{
NE <- getDataBBS(file = "Nebrask.zip")
}

}
