% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conditions.R
\name{get_conditions}
\alias{get_conditions}
\title{Download and import the annual conditions per route  via FTP (USGS BBS)}
\usage{
get_conditions(
  conds.url = "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/Weather.zip",
  data.dir = here::here("raw-data/"),
  active.only = FALSE,
  overwrite = NULL
)
}
\arguments{
\item{conds.url}{URL for location of the routes.zip folder.}

\item{data.dir}{Location for storing the routes.zip and unzipped file(s). We recommend storing this in the raw-data directory.}

\item{active.only}{Logical. Default = FALSE. If true, keeps only the active routes.}

\item{overwrite}{Logical. Default = NULL. If true, will download and overwrite the existing file "Weather.zip" in data.dir.}
}
\value{
A dataframe containing route-level information.
}
\description{
This function downloads the 'weather.zip' file from the BBS FTP server. This data contains information on local weather conditions during each survey. Also includes auxiliary information about the survey methods, dates of surveys, and start and end times. This data will typically be incorporated into detection probability functions.
}
\examples{
# download and import the BBS route information
\dontrun{
conditions.df <- get_conditions()
}

}
