% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credibility_trends.R
\name{get_credibility_trends}
\alias{get_credibility_trends}
\title{Get data credibility and trends from the BBS analysis results}
\usage{
get_credibility_trends(
 
    url = "https://www.mbr-pwrc.usgs.gov/cgi-bin/atlasa15.pl?FLA&2&15&csrfmiddlewaretoken=3YKakk7LxT2ki6NSpl4mstudYCqdW02C"
)
}
\arguments{
\item{url}{Web address of the region or state for which species' population trends and credibility ratings are to be scraped. Default example is 'Kansas,USA'.}
}
\description{
The BBS provides regional and state-wide species population trend estimates. They also include a 'data credibility' rating (low, medium, or high deficiencies) for each species-region combination. This function uses web scraping to capture population trends and their credibility ratings.
}
