% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unzip.R
\name{get_unzip}
\alias{get_unzip}
\title{Downloads and unzips a zip archive to file from the USGS BBS FTP server}
\usage{
get_unzip(ZipName, FileName)
}
\arguments{
\item{ZipName}{file to download}

\item{FileName}{file to unzip to}
}
\value{
A dataframe
}
\description{
This function downloads the BBS data from the USGS FTP server to a temporary file, decompresses the .zip files, and imports the uncompressed .csv file into R. This function was adapted from the function \href{https://github.com/oharar/rBBS/blob/master/R/GetUnzip.R}{rBBS::GetUnzip()}.
}
\details{
Used internally. If ZipName begins with 'http' or 'ftp', then download and unzip to
Filename and return as a dataframe. Otherwise, unzip the file specified by ZipName and import into R as a data.frame.
}
