% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_bbs.R
\name{import_bbs}
\alias{import_bbs}
\title{Import BBS data into R from file.}
\usage{
import_bbs(
  data.dir = paste0(getwd(), "/raw-data/"),
  state.names = NULL,
  overwrite.routes = TRUE
)
}
\arguments{
\item{data.dir}{Location of the compressed (.zip) route-level files (one per region/states) to be imported into R. Files will also be unzipped to this directory as ".csv".}

\item{state.names}{Vector of state names Default = NULL (all states). See column 'State' in data("region_codes").}

\item{overwrite.routes}{Logical. Default TRUE will download and overwrite any existing files named "routes.csv" in data.dir.}
}
\description{
Import compressed (.zip) BBS data from file, combine all regions/states into a single data.frame object and load into environment.
}
