% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_species_list.R
\name{import_species_list}
\alias{import_species_list}
\title{Import the species list comprising names and numeric identifiers.}
\usage{
import_species_list(sb_dir)
}
\arguments{
\item{sb_dir}{Directory for the ScienceBase (sb) item.}
}
\description{
Imports the file specieslist.txt into R env
}
