% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{region_codes}
\alias{region_codes}
\title{Region Codes Lookup Table}
\format{
An object of class \code{data.frame} with 96 rows and 7 columns.
}
\source{
\url{https://www.sciencebase.gov/catalog/item/52b1dfa8e4b0d9b325230cd9}
}
\usage{
region_codes
}
\description{
Country and state codes used by BBS combined with ISO-3166-2 and ISO-a2
}
\details{
\describe{
\item{CountryNum}{alpha-country codes used by the BBS. (840=United States of America; 484=Mexico; 124=Canada)}
\item{StateNum}{state numbers associated with each State and CountryNum, assigned by the BBS}
\item{State}{proper name of the state or province as provided by BBS office}
\item{ISO_a2}{ISO 2-character code for Country}
\item{ISO_3166-2}{ISO character code for Country and State/Territory/Province}
\item{State}{English name associated with ISO_3166-2. Names appearing in the BBS region codes list with accented characters were munged by development team slightly.}
\item{name_fr}{French name associated with ISO_3166-2. Names appearing in the BBS region codes list with accented characters were munged by development team slightly.}
\item{name_es}{Spanish (Espanol) name associated with ISO_3166-2. Names appearing in the BBS region codes list with accented characters were munged by development team slightly.}
}
}
\keyword{datasets}
