% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_codes}
\alias{region_codes}
\title{Country and state codes and associated zip filenames}
\format{
A data frame:
\describe{
\item{CountryNum}{Integer. BBS country codes associated with United States-840, Mexico-484, and Canada-124}
\item{StateNum}{Integer. BBS state numbers associated with each 'StateNum' and 'CountryNum'}
\item{State}{Character. The proper name of the US or Mexican state or Canadian province.}
\item{zip_states}{Character. The filename of the .zip compressed BBS count data located on the FTP server. These files do not exist for Mexican states, so == NA}
}
}
\source{
\url{https://www.sciencebase.gov/catalog/item/52b1dfa8e4b0d9b325230cd9}
}
\usage{
region_codes
}
\description{
Country and state codes and associated zip filenames
}
\keyword{datasets}
