% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_codes}
\alias{region_codes}
\title{Country and state codes and associated zip filenames}
\format{
A data frame:
\describe{
\item{CountryNum}{BBS country codes associated with United States-840, Mexico-484, and Canada-124}
\item{StateNum}{BBS state numbers associated with each 'State' and 'CountryNum'}
\item{State}{The proper name of the US or Mexican state or Canadian province}
\item{zip_states}{The filename of the .zip compressed BBS count data located on the FTP server. These files do not exist for Mexican states, so == NA}
}
}
\source{
\url{ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/RegionCodes.txt}
}
\usage{
region_codes
}
\description{
Country and state codes and associated zip filenames
}
\keyword{datasets}
