% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sauer_results}
\alias{sauer_results}
\title{Contains the most recent release of the BBS observations dataset.
The data object is a list, wherein the list names (names(sauer_results)) correspond with the .csv filenames provided in ScienceBase; call \code{bbsAssistant::sb_items} for more information.}
\format{
A list containing a varying number of data frames associated with the most recent release of analysis results by Sauer et al. Please visit the \code{source} for more information, including metadata which describes each results file.
\describe{
NOTE: THE COLUMNS HERE ARE DESCRIBED IN GREATER DETAIL AT THE SCIENCEBASE ITEM (see "sb_items" hyperlink for relevant sb_id).
NOTE ALSO: NOT ALL COLUMNS ARE DESCRIBE HERE. CURRENTLY, ONLY THE COLUMNS ASSOCIATED WITH THE 2020 RELEASE ARE DESCRIBED.
@details The object, sauer_results, comprises (currently) 4 data frames, each containing the results of different analyses as described in original metadata.
\itemize{ # the list
\item{@param BBS_1966-2018_core_best_trend}{...}
\item{@param BBS_1993-2018_expanded_trend_best}{ ... }
\item{@param inde_best_1993-2018_expanded}{ ... }
\item{@param inde_best_1966-2018_core}{ ... }
}
}
\source{
\url{https://www.sciencebase.gov/catalog/item/5ea1e02c82cefae35a16ebc4}
}
\usage{
sauer_results
}
\description{
Contains the most recent release of the BBS observations dataset.
The data object is a list, wherein the list names (names(sauer_results)) correspond with the .csv filenames provided in ScienceBase; call \code{bbsAssistant::sb_items} for more information.
}
\keyword{datasets}
