% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_list}
\alias{species_list}
\title{Species List Associated with Most Recent BBS Dataset Release}
\format{
A \code{data.frame} with 756 rows and 9 variables
\describe{
\item{Seq}{row identifier used by BBS}
\item{AOU}{species-level numeric code used by BBS}
\item{English_Common_Name}{Common name in English}
\item{French_Common_Name}{Common name in French}
\item{Spanish_Common_Name}{Common name in Spanish}
\item{ORDER}{taxonomic order}
\item{Family}{taxonomic family}
\item{Genus}{taxonomic genus}
\item{Species}{taxonomic species epithet}
}
\source{
\url{https://www.sciencebase.gov/catalog/item/52b1dfa8e4b0d9b325230cd9}
}
\usage{
species_list
}
\description{
A species list is provided by the BBS office in ScienceBase data releases.See the ScienceBase catalog for metadata regarding taxonomic classifications.
}
\keyword{datasets}
