% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_list}
\alias{species_list}
\title{Basic taxonomy (AOU, NACC) merged with the most recent release of the BBS observations dataset.}
\format{
A data frame .
\describe{
\item{AOU}{Numeric AOU code.}
\item{NACC}
}
}
\source{
Most recent version of NACC at \url{http://checklist.aou.org/taxa.csv?type=charset\%3Dutf-8\%3Bsubspecies\%3Dno\%3B}

Most recent version of AOU at \url{http://checklist.aou.org/taxa.csv?type=charset\%3Dutf-8\%3Bsubspecies\%3Dno\%3B}

Most recent version of species list associated with BBS data at SpeciesList.txt at \url{http://checklist.aou.org/taxa.csv?type=charset\%3Dutf-8\%3Bsubspecies\%3Dno\%3B"}
}
\usage{
species_list
}
\description{
Basic taxonomy (AOU, NACC) merged with the most recent release of the BBS observations dataset.
}
\keyword{datasets}
