% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_list}
\alias{species_list}
\title{Taxonomic information from the American Ornithological Society's AOU code.}
\format{
A data frame with 2158 rows and 17 variables.
\describe{
\item{AOU}{Numeric AOU code.}
\item{avibase.id}{Identifier for Avibase database. Source \url{http://avibase.bsc-eoc.org/}}
\item{annotation}{...}
\item{rank}{...}
\item{common_name}{English common name.}
\item{order}{Phylogenetic order.}
\item{family}{Phylogenetic family}
\item{subfamily}{Phylogenetic sub-family}
\item{genus}{Phylogenetic genera}
\item{species}{Phylogenetic species epithet}
\item{status_accidental}{...}
\item{status_hawaiin}{}
\item{status_introduced}{...}
\item{status_misplaced}{...}
\item{status_extinct}{...}
\item{status_nonbreeding}{...}
}
}
\source{
Most recent version of AOU at \url{http://checklist.aou.org/taxa.csv?type=charset\%3Dutf-8\%3Bsubspecies\%3Dno\%3B}

Most recent version of species list associated with BBS data at SpeciesList.txt at \url{http://checklist.aou.org/taxa.csv?type=charset\%3Dutf-8\%3Bsubspecies\%3Dno\%3B"}
}
\usage{
species_list
}
\description{
Taxonomic information from the American Ornithological Society's AOU code.
}
\keyword{datasets}
